Rewrite summarize-telemetry.js to not crash server
As the number of log files increases (one is created per day), the
amount of RAM required to run this script (which runs daily in a
cron-job) as increased.

The Promise.all() call just kicks off all file reads at the same time,
which uses a ton of resources. The webtorent.io server is configured to
restart when 100% RAM is used to prevent a common situation with
limited-RAM VPS servers where the swap file is used and thrashing
occurs, basically locking up the server.

I've noticed the server restarting itself once per day whenever this
cron job runs.

This rewrite just removes Promises (sorry!) and uses run-parallel-limit
to limit the concurrency to 5 tasks at a time. There's probably a
promise way to do this, but didn't want to figure it out right now.